Add basic tests
The strategy so far is to create a dummy sails application, into which
the hook under test can be loaded.

Barrels is used to load test fixture data into a sails-memory database.

`sails.request` is used to make a fake client request to get the model data,
and the format of the response body is asserted to be as expected for jsonapi.

To run, `node <path/to/test.js>` can be used.